Changed axis type from OrderedDict to nested list (dict)
Was previously storing axes as an orderd dictionary, but forsaw that this was going to be annoying if we want to collapse or create new axes.  Tried a named tuple, but they can't be changed so that's no good. Settled on:

self.axes = [    {'name':name, 'axis':axis  ]

where axis is still an astropy quantity (w untis)