Changed bdot offset calculation (Issue #6)
Bdot offset was previously removed by linearly detrending the entire signal.

Now, the offset is calculated as the mean of a small number of points at the start of the signal, which is then subtracted from the whole signal. A keyword "offset_range" has been added to set this range. Default values are (0,100), which should work even for the HRR setup. Longer times are better when available.