Tried to do HF helmholtz calibration using Everson2009 Eq. 5: no success
The Helmholtz coil really does fail at frequencies above 100 kHz.  Eq. 7 is the wrong equation to use here, because it is a 1st order expansion. Fitting with Eq. 5 works up to 100 kHz, but then clearly deviates. It seems that the straight wire antenna really is necessary for HF calibration to determin tau.