Introduced "trange" keyword to load/lapd.py
Added the capability to load only a segment of the data (in time) into a full file.
This is useful because we often digitize everything we can to avoid missing anything, but in practice only end up using a small part of the data. This keyword can then be used to limit file sizes and processing times.