Bugfix: t0indarr for data taken on slow DAQ
The t0 index determined on the tdiode datafiles was previously used to determine t0 on all other datasets. This was in error for data taken at a different clock rate than the tdiode. For example, the t0 calculated from a tdiode on the 1.25 GHz fast DAQ is incorrect for a bdot on the 100 MHz slow DAQ. A correction has been introduced in all of the affected processing routines, and a "t0arr" as been added to the tdiode file to help identify similar problems in the future.