Tdiode calcT0ind bugfix
On bad shots, the maximum of the Tdidoe array can be in the first several elements, making the array in this line too short to calculate the gradient. This try/except statement catches the resulting ValueError and just puts 0 in the array: this shot should be ignored anyway so the value doesn't really matter.