Added Error Message to HRR
If no probe motion is actually included in the run but the "_pos_resource" keywords etc. are set in the csv, the program throws an error because it looks for the position information and can't find it. The Phoenix Termianl HRR doesn't write position information to the HDF unless the probe actually moves. Right now I think the best way to handle this is to set those fields to NA in the csv if the probe does not move, but we can discuss other ways of handling it. I added a more informative error message that should help if this happens again.