Improved error handling in langmuir vsweep routine
Moved error handling into the vsweep_fit function and made it more strict. Several cases that indicate badly shaped data result in a default value being returned, which can now be set with the variable fail_val. Moved the calculation of dimensional quantities into the function to reduce duplication in the main program.

Also added a statement to delete a previous version of the destination files if one exists, and removed an older attempt at doing this. Not sure why this way wasn't done before: make sure it works on OSX.