Added fit error output to langmuir.py vsweep_fit
This routine now outputs an additional variable "ebars" that contains fitting errors (or, if the fit fails, the fail_val).

In theory, it should be possible to propagte these errors through the calculations and turn them into real dimensional error bars on the predicted values. However, doing this through the equation for V_pp is difficult. Someone should do that.

The errors (in arb fitting units) in the ebar array correspond to:

#1,2 : slope and offset of esat
#3,4: amplitude and offset of expoenntial region
#5: temperature